Stop activity if we cant start node
Currently when receiving a push notification while the app has not
started (on android), the whole app is woken up. If autologin is enabled, we will try
to start the node, but will fail has no activity has been attached.
In such cases we try for 10s and quit the app if not successful.

Signed-off-by: Andrea Maria Piana <andrea.maria.piana@gmail.com>