Change recover to access
There are some instances on the keycard screens that read `Recover key` rather than `Access key`, this is dated language so I've changed it. There are a few other instances of the word 'recover' in relation to keycard that may also need to be changed, but I'm not sure what context they are in.

Signed-off-by: Andrey Shovkoplyas <motor4ik@gmail.com>